+++ src/frontend/web_application/src/scenes/UserProfile/components/ProfileForm/index.jsx
@@ -1,5 +1,0 @@
-import { withI18n } from '@lingui/react';
-import { compose } from 'redux';
-import Presenter from './presenter';
-
-export default compose(withI18n())(Presenter);