+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Fixed
+
+- Unable to to delete account.
+
 ## [0.24.5] 2021-08-20
 
 ### Fixed